+++ components/google-cloud/google_cloud_pipeline_components/experimental/remote/gcp_launcher/Dockerfile
@@ -1,10 +1,10 @@
-# Copyright 2021 The Kubeflow Authors. All Rights Reserved.
+# Copyright 2021 The Kubeflow Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+# http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,15 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# Base image to use for this docker
-FROM gcr.io/google-appengine/python:latest
+FROM python:3.7-slim
 
-WORKDIR /root
+RUN pip install -U google-cloud-aiplatform
+RUN pip install -U google-cloud-storage
 
-# Upgrade pip to latest
-RUN pip3 install --upgrade pip
-
-# TODO() change from using PR commit id to release 
 RUN pip3 install "git+https://github.com/kubeflow/pipelines.git#egg=google-cloud-pipeline-components&subdirectory=components/google-cloud"
 
-ENTRYPOINT ["python3","-m","google_cloud_pipeline_components.aiplatform.remote_runner"]
+ENTRYPOINT ["python", "-u", "-m", "google_cloud_pipeline_components.experimental.remote.gcp_launcher.launcher"]
