+++ components/google-cloud/google_cloud_pipeline_components/experimental/remote/gcp_launcher/__init__.py
@@ -1,22 +1,14 @@
-# Copyright 2021 The Kubeflow Authors
+# Copyright 2021 The Kubeflow Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-# http://www.apache.org/licenses/LICENSE-2.0
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-FROM python:3.7-slim
-
-RUN pip install -U google-cloud-aiplatform
-RUN pip install -U google-cloud-storage
-
-COPY google_cloud_pipeline_components/experimental/gcp_launcher .
-
-ENTRYPOINT ["python", "-u", "-m", "launcher"]
+"""GCP Launcher Modules for google cloud components."""
