+++ components/google-cloud/google_cloud_pipeline_components/experimental/remote/gcp_launcher/cloudbuild.yaml
@@ -1,0 +1,6 @@
+steps:
+- name: 'gcr.io/kaniko-project/executor:latest'
+  args: 
+  - --destination=gcr.io/$PROJECT_ID/gcp-launcher:latest
+  - --cache=true
+  - --cache-ttl=12h
