+++ components/google-cloud/tox.ini
@@ -33,6 +33,9 @@ commands =
    flake8 . --count --select=E9,F63,F7,F82,W --show-source --statistics
    py.test --cov=google_cloud_pipeline_components --cov-append --cov-report=term-missing -vvv -s {posargs}
 
+[coverage:report]
+skip_empty = true
+
 [testenv:clean]
 deps = coverage
 skip_install = true
